internal/glutil: make [Un]ConvertRect operate on standard ints.

+++ internal/glutil/rect.go
@@ -6,23 +6,23 @@ package glutil
 
 import "image"
 
-func ConvertRect(rect, bounds image.Rectangle) (x, y, width, height int32) {
+func ConvertRect(rect, bounds image.Rectangle) (x, y, width, height int) {
 	// We must flip the Y axis because image.Rectangle uses top-left as
 	// the origin but OpenGL uses bottom-left as the origin.
-	y = int32(bounds.Dy() - (rect.Min.Y + rect.Dy())) // bottom
-	height = int32(rect.Dy())                         // top
+	y = bounds.Dy() - (rect.Min.Y + rect.Dy()) // bottom
+	height = rect.Dy()                         // top
 
-	x = int32(rect.Min.X)
-	width = int32(rect.Dx())
+	x = rect.Min.X
+	width = rect.Dx()
 	return
 }
 
-func UnconvertRect(bounds image.Rectangle, x, y, width, height int32) (rect image.Rectangle) {
+func UnconvertRect(bounds image.Rectangle, x, y, width, height int) (rect image.Rectangle) {
 	// We must unflip the Y axis because image.Rectangle uses top-left as
 	// the origin but OpenGL uses bottom-left as the origin.
-	x0 := int(x)
-	x1 := int(x + width)
-	y0 := bounds.Dy() - int(y+height)
-	y1 := y0 + int(height)
+	x0 := x
+	x1 := x + width
+	y0 := bounds.Dy() - (y + height)
+	y1 := y0 + height
 	return image.Rect(x0, y0, x1, y1)
 }
